refactor(types): move PerTurnContext to src/types/agent.ts
CodeRabbit on PR #784: importing PerTurnContext from a UI module
(`agent-view-tool-followup.ts`) into the deliberately UI-agnostic
`AgentLoop` is a layering violation — the loop should not reach
through to the agent-view tree for any of its types.

Promote the type to the existing `src/types/agent.ts` module (the
established home for agent-level types). `agent-loop.ts` now imports
it from there alongside `ChatSession`. The function helpers
(buildFollowUpRequest etc.) still live in agent-view-tool-followup.ts
— relocating those is a separate, larger refactor outside this fix.

The followup module re-exports the type so existing UI-internal
imports (`agent-view-tools.ts`) keep working unchanged.

No behavior change; tests + format-check still green.
